chore: conversation fixes
