+++ src/hooks/useClientIntegration.ts
@@ -14,7 +14,6 @@ const useClientIntegration = () => {
   const client = isInPlayer ? ClientIntegrations.INPLAYER : ClientIntegrations.CLEENG;
   const clientId = isInPlayer ? integrations?.inplayer?.clientId : integrations?.cleeng?.id;
 
-  // TODO: More data will follow. Example: access fees will be fetched and returned once user is auth.
   return {
     integration: integrations,
     client,
