+++ types/inplayer.d.ts
@@ -1,0 +1,4 @@
+export type InPlayerAuthData = {
+  access_token: string;
+  expires?: number;
+};
